Move ping test service to testutils for reuse
