+++ pkg/testutils/grpc/ping/doc.go
@@ -1,4 +1,4 @@
-// Package test provides a grpc service for testing
+// Package ping provides a grpc service for testing
 // The service allows controlling the response returned by the server
 // including the generation of specific errors.
-package test
+package ping
