+++ pkg/testutils/grpc/ping/ping_server_test.go
@@ -1,4 +1,4 @@
-package test
+package ping
 
 import (
 	context "context"
